LineSegmentationAdded
Added line segmentation to ParseAndModifyKMZ.py, and added Batch file to run the script. The script is nearly complete. I noticed a small error where the offset is larger than intended, and another where there are switchbacks at a few 90* or nearly 90* angles in the line when the segments are drawn. I will have to get to these tomorrow as I just saw it's after midnight!