SwitchbacksAndOvershootsFixed
As noted in the README, I fixed the switchback/overshoot problem. The final vertices leading up to the right angles (those overlapping the point features) had been getting skipped because there was a loophole in the conditional statements that didn't recognize and account for vertices with a measured distance of 0 to the next point feature.